fix: rename routables to routetables
